The list of tests is now a list instead of a dict in the output JSON.
This change reflects the internal refactoring of the correponding data
structure.
A dict was unsuitable for two reasons:

1. it did not support tests with the same title (not recommended, but
definitely possible)
2. it needed the extra field "order" to keep track of the ordering of
the tests in the corresponding test file

Now the 'tests' field in the output JSON is a list of dicts, instead of
a dict. This maintain the ordering and allows multiple tests with the
same name. The test name is stored in the 'title' field.